refactor(web): decompose AuthContext responsibilities
- Extract backend readiness poller
- Extract authentication helpers
- Simplify AuthContext provider from 548 to 398 LOC
