+++ apps/web/src/context/AuthContext.tsx
@@ -14,11 +14,17 @@ import {
 import { useRouter } from "next/navigation";
 import type { User } from "@/types/User";
 
-import { apiClient } from "@/lib/api/client";
-import { isAPIError } from "@/lib/api/APIError";
 import { normalizeError } from "@/lib/api/normalizeError";
 import { authApi } from "@/lib/api/auth";
 import logger from "@/lib/logger";
+
+import {
+  AUTH_SESSION_STORAGE_KEY,
+  isValidUser,
+  isBenignLogoutError,
+  replaceToHomeSafely,
+} from "./auth/authHelpers";
+import { useBackendReadyPoller } from "./auth/useBackendReadyPoller";
 
 /* -------------------------------------------------------------------------- */
 /* Types                                                                      */
@@ -28,8 +34,6 @@ export type AuthStatus =
   | "loading"
   | "unauthenticated"
   | "authenticated";
-
-const AUTH_SESSION_STORAGE_KEY = "esparex_user_session";
 
 interface AuthContextType {
   user: User | null;
@@ -53,51 +57,8 @@ interface BackendReadyContextType {
 const AuthContext =
   createContext<AuthContextType | undefined>(undefined);
 
-// Separated from AuthContext so that the backend cold-start transition
-// (false → true) only re-renders the one component that needs it (useOtpFlow),
-// not every component that calls useAuth().
 const BackendReadyContext =
   createContext<BackendReadyContextType | undefined>(undefined);
-
-/* -------------------------------------------------------------------------- */
-/* Utils                                                                      */
-/* -------------------------------------------------------------------------- */
-
-function isValidUser(data: unknown): data is User {
-  if (!data || typeof data !== "object") return false;
-  const candidate = data as Partial<User>;
-
-  return (
-    typeof candidate.id === "string" &&
-    typeof candidate.mobile === "string" &&
-    typeof candidate.role === "string"
-  );
-}
-
-function isBenignLogoutError(error: unknown): boolean {
-  if (!isAPIError(error)) return false;
-
-  if (error.status !== 401) return false;
-
-  const backendMessage = String(
-    error.context?.backendErrorMessage ??
-      (typeof error.details === "object" && error.details !== null && "error" in error.details
-        ? (error.details as { error?: unknown }).error
-        : "") ??
-      ""
-  ).toLowerCase();
-
-  return backendMessage.includes("no token") || error.message.toLowerCase().includes("no token");
-}
-
-function replaceToHomeSafely(router: ReturnType<typeof useRouter>) {
-  if (typeof window !== "undefined") {
-    window.location.replace("/");
-    return;
-  }
-
-  void router.push("/");
-}
 
 /* -------------------------------------------------------------------------- */
 /* Provider                                                                   */
@@ -111,20 +72,15 @@ export function AuthProvider({
   initialHasAuthCookie?: boolean;
 }) {
   const router = useRouter();
-  const [user, setUser] =
-    useState<User | null>(null);
+  const [user, setUser] = useState<User | null>(null);
+  const [error, setError] = useState<Error | null>(null);
 
+  const { backendReady, setBackendReady, hasAuthHint, setHasAuthHint } =
+    useBackendReadyPoller(initialHasAuthCookie);
 
-  const [error, setError] =
-    useState<Error | null>(null);
-
-  const [backendReady, setBackendReady] =
-    useState(false);
-  const [hasAuthHint, setHasAuthHint] =
-    useState(initialHasAuthCookie);
-
-  const [status, setStatus] =
-    useState<AuthStatus>(initialHasAuthCookie ? "loading" : "unauthenticated");
+  const [status, setStatus] = useState<AuthStatus>(
+    initialHasAuthCookie ? "loading" : "unauthenticated"
+  );
 
   const fetchingRef = useRef(false);
   const authBannerShownRef = useRef(false);
@@ -142,15 +98,11 @@ export function AuthProvider({
 
     if (!backendReady) {
       if (
-        process.env.NEXT_PUBLIC_LOCAL_DEV_AUTH ===
-        "true" &&
+        process.env.NEXT_PUBLIC_LOCAL_DEV_AUTH === "true" &&
         process.env.NODE_ENV !== "production"
       ) {
-        logger.warn(
-          "⚠️ LOCAL DEV AUTH ENABLED — DO NOT DEPLOY"
-        );
+        logger.warn("⚠️ LOCAL DEV AUTH ENABLED — DO NOT DEPLOY");
 
-        /* ✅ FIXED DEV USER */
         const devUser: User = {
           id: "local-dev-user",
           name: "Local Dev User",
@@ -183,9 +135,6 @@ export function AuthProvider({
 
     try {
       const response = await authApi.me({ silent: true });
-
-      /* Normalize API Shapes */
-      // authApi.me() returns { success: boolean, user: User }
       const rawUser = response.user;
 
       if (isValidUser(rawUser)) {
@@ -207,18 +156,16 @@ export function AuthProvider({
     } catch (rawError: unknown) {
       const err = normalizeError(rawError);
 
-      /* Auth failures are valid states */
       if (
         err.response?.status === 401 ||
         err.response?.status === 403
       ) {
-        // One-time stale-session cleanup for invalid/blacklisted HttpOnly cookies.
         if (!staleSessionCleanupRef.current) {
           staleSessionCleanupRef.current = true;
           try {
             await authApi.logout();
           } catch {
-            // Ignore cleanup failures; UI still transitions to unauthenticated.
+            // Ignore cleanup failures
           }
         }
 
@@ -240,14 +187,11 @@ export function AuthProvider({
         return;
       }
 
-      /* Network error — backend is sleeping (Render free tier) or unreachable.
-         Do NOT treat this as a session expiry. Stay in "loading" and retry so
-         the user isn't kicked to the login page just because the backend is cold. */
       if (!err.response) {
         const MAX_NETWORK_RETRIES = 3;
         if (networkRetryCountRef.current < MAX_NETWORK_RETRIES) {
           networkRetryCountRef.current += 1;
-          const delay = networkRetryCountRef.current * 5_000; // 5s, 10s, 15s
+          const delay = networkRetryCountRef.current * 5_000;
           if (process.env.NODE_ENV === "development") {
             logger.warn(`[Auth] Network error — retrying in ${delay / 1000}s (attempt ${networkRetryCountRef.current}/${MAX_NETWORK_RETRIES})`);
           }
@@ -255,18 +199,14 @@ export function AuthProvider({
             fetchingRef.current = false;
             void doFetch();
           }, delay);
-          // Stay in "loading" — withGuard shows null, no login redirect
           return;
         }
-        // Give up after MAX_NETWORK_RETRIES — backend is genuinely down
         networkRetryCountRef.current = 0;
         setUser(null);
         setStatus("unauthenticated");
-        // Preserve hasAuthHint so next page load re-attempts without requiring re-login
         return;
       }
 
-      /* Rate limit / boot noise */
       if (
         err.isExpected ||
         err.response?.status === 429
@@ -278,19 +218,13 @@ export function AuthProvider({
         return;
       }
 
-      if (
-        process.env.NODE_ENV === "development"
-      ) {
-        logger.error(
-          "[Auth] Fetch failed:",
-          err.message
-        );
+      if (process.env.NODE_ENV === "development") {
+        logger.error("[Auth] Fetch failed:", err.message);
       }
 
       setError(
         new Error(
-          err.message ||
-          "Authentication failed"
+          err.message || "Authentication failed"
         )
       );
 
@@ -300,84 +234,7 @@ export function AuthProvider({
     } finally {
       fetchingRef.current = false;
     }
-  }, [backendReady, router]);
-
-  /* ------------------------------------------------------------------------ */
-  /* Backend Health Check                                                     */
-  /* ------------------------------------------------------------------------ */
-
-  useEffect(() => {
-    if (initialHasAuthCookie || typeof window === "undefined") return;
-
-    if (localStorage.getItem(AUTH_SESSION_STORAGE_KEY) === "1") {
-      setHasAuthHint(true);
-      setStatus((current) => (current === "unauthenticated" ? "loading" : current));
-    }
-  }, [initialHasAuthCookie]);
-
-  useEffect(() => {
-    let mounted = true;
-
-    if (
-      process.env.NEXT_PUBLIC_LOCAL_DEV_AUTH ===
-      "true" &&
-      process.env.NODE_ENV !== "production"
-    ) {
-      setBackendReady(true);
-      return;
-    }
-
-    /**
-     * Exponential backoff with jitter.
-     *
-     * Schedule:
-     *   Attempt 0 →  2 s  + jitter (0–1 s)
-     *   Attempt 1 →  4 s  + jitter
-     *   Attempt 2 →  8 s  + jitter
-     *   Attempt 3 → 16 s  + jitter
-     *   Attempt 4+→ 30 s  + jitter (cap)
-     *
-     * During a sustained outage this reduces health-check traffic from
-     * ~30 req/min (fixed 2 s loop) to ≤2 req/min once fully backed off.
-     * Counter resets to 0 immediately on a successful health response.
-     */
-    const BASE_DELAY_MS = 2_000;
-    const MAX_DELAY_MS = 30_000;
-    let retryAttempt = 0;
-
-    const waitForBackend = async () => {
-      try {
-        const ok = await apiClient.checkHealth();
-
-        if (mounted && ok) {
-          retryAttempt = 0;
-          setBackendReady(true);
-        } else if (mounted) {
-          const jitter = Math.random() * 1_000;
-          const delay =
-            Math.min(BASE_DELAY_MS * Math.pow(2, retryAttempt), MAX_DELAY_MS) +
-            jitter;
-          retryAttempt += 1;
-          setTimeout(waitForBackend, delay);
-        }
-      } catch {
-        if (mounted) {
-          const jitter = Math.random() * 1_000;
-          const delay =
-            Math.min(BASE_DELAY_MS * Math.pow(2, retryAttempt), MAX_DELAY_MS) +
-            jitter;
-          retryAttempt += 1;
-          setTimeout(waitForBackend, delay);
-        }
-      }
-    };
-
-    waitForBackend();
-
-    return () => {
-      mounted = false;
-    };
-  }, []);
+  }, [backendReady, router, setBackendReady, setHasAuthHint]);
 
   /* ------------------------------------------------------------------------ */
   /* Session Sync                                                             */
@@ -397,7 +254,6 @@ export function AuthProvider({
       return;
     }
 
-    // Force a fetch if a business update event occurs
     const handleAuthUpdate = () => {
       if (!pathname.startsWith("/admin")) {
         fetchUser();
@@ -449,7 +305,7 @@ export function AuthProvider({
       }
       wasAuthenticatedRef.current = true;
     },
-    []
+    [setHasAuthHint]
   );
 
   /* ------------------------------------------------------------------------ */
@@ -486,7 +342,7 @@ export function AuthProvider({
       authBannerShownRef.current = false;
       staleSessionCleanupRef.current = false;
     }
-  }, []);
+  }, [setHasAuthHint]);
 
   /* ------------------------------------------------------------------------ */
   /* Provider                                                                 */
@@ -520,7 +376,7 @@ export function AuthProvider({
 }
 
 /* -------------------------------------------------------------------------- */
-/* Hook                                                                       */
+/* Hooks                                                                      */
 /* -------------------------------------------------------------------------- */
 
 export function useAuth() {
@@ -535,11 +391,6 @@ export function useAuth() {
   return ctx;
 }
 
-/**
- * useBackendReady — subscribe only to backendReady.
- * Only this hook re-renders on the backend cold-start transition;
- * all other useAuth() consumers are unaffected.
- */
 export function useBackendReady(): boolean {
   const ctx = useContext(BackendReadyContext);
   if (!ctx) throw new Error("useBackendReady must be used within AuthProvider");
